api/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.118.0
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
jinja2==3.1.6
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==11.3.0
pydantic==2.11.10
pydantic-core==2.33.2
python-multipart==0.0.20
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
triton==3.4.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
